feat(@eurora/desktop): add code for dynamic registering and adding of extensions
